Move ThemeGenerator to common directory.
